[#20] Design : 1대1 채팅방 퍼블리싱 채팅방 클릭 전, 클릭 후 링크 구분
