+++ src/app/(chats)/chats/layout.module.css
@@ -1,0 +1,134 @@
+.container {
+  display: flex;
+  justify-content: center;
+  height: 100%;
+	
+}
+
+.formContainer {
+  width: 100%;
+	height: 100%;
+  border: 1px solid rgb(223, 223, 223);
+	
+}
+
+.container2 {
+	height: 100%;
+	display:flex;
+	
+}
+.roomList {
+	width: 35%;
+	padding-top: 30px;
+	max-height: 100%;
+	overflow: auto;
+	border-right: 1px solid lightgray;
+}
+.roomCount {
+
+	font-size: larger;
+	font-weight: bold;
+	padding-bottom: 30px;
+	padding-left: 20px;
+	border-bottom: 1px solid lightgray;
+}
+.list {
+	display: flex;
+	padding-top: 15px;
+	padding-bottom: 15px;
+	padding-right: 5px;
+	padding-left: 5px;
+	
+}
+.list:hover {
+	background-color: rgb(220, 220, 220);
+}
+
+.profileImg {
+	width: 40px;
+	height: 40px;
+	border-radius: 50%;
+	border: 1px solid lightgray;
+}
+
+.productImg {
+	width: 35px;
+	height: 35px;
+	border-radius: 50%;
+	border: 1px solid lightgray;
+	
+}
+
+.shortcut {
+	width: 100%;
+	padding-left: 7px;
+	padding-top: 3px;		
+}
+
+.name {
+	font-weight: bold;
+	padding-bottom: 5px;
+}
+
+.msgArea {
+	display: flex;
+	justify-content: space-between;
+}
+
+.lastMsg {
+	font-size: small;
+	color: gray;
+	width: 300%
+}
+
+.lastMsgDate {
+	font-size: small;
+	color: gray;
+	width: 100%;
+	text-align:right;
+	padding-right: 3px;
+}
+
+.chatRoom {
+	padding-top: 20px;
+	max-height: 100%;
+	width: 65%;
+	overflow: auto;
+	
+}
+
+.noChatRoom {
+	height: 100%;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	
+}
+
+.height {
+	height: 90%
+}
+
+.noRoom svg {
+	font-size: 75px;
+}
+
+.noChat svg {
+	font-size: 75px;
+}
+
+.noChatFont {
+	font-size: larger;
+	text-align: center;
+	font-weight: bold;
+}
+
+.chatIcon {
+	display: flex;
+	justify-content: center;
+}
+
+.noChatFontSmall {
+	font-size: medium;
+	text-align: center;
+}
