+++ src/app/(chats)/chats/page.tsx
@@ -1,0 +1,5 @@
+export default function Page() {
+	return (
+		<></>
+	);
+}
