--- src/app/(talks)/talks/layout.module.css
@@ -1,13 +1,0 @@
-.container {
-  display: flex;
-  justify-content: center;
-  height: 100%;
-	
-}
-
-.formContainer {
-  width: 100%;
-	height: 100%;
-  border: 1px solid rgb(223, 223, 223);
-	
-}
