--- src/app/(talks)/talks/layout.tsx
@@ -1,14 +1,0 @@
-import styles from "./layout.module.css"
-export default function TalksLayOut({
-	children,
-}: Readonly<{
-	children: React.ReactNode;
-}>) {
-	return (
-		<div className={styles.container}>
-			<div className={styles.formContainer}>
-				{children}
-			</div>
-		</div>
-	)
-}
